Add files via upload
This is the original dataset that I worked with for this project. I accessed the data via Kaggle.com from Aminat Lawal, who scraped the data from the Wikipedia page "List of ended Netflix original programming". There were 1,353 entries, with each listing the name of the show, the premiere date, the finale date, the number of seasons the show lasted, and whether the show was cancelled after its first release.