Add files via upload
This is how the dataset looked after it was cleaned and before it was ready for analysis.
Before starting the project, I served as the theoretical stakeholder and asked a few questions I wanted to take away from this analysis. I wanted to determine how dramas compared to comedies in regard to durational success, how many shows premiered and ended for each year, and if certain seasonal quarters were more likely to get a show to premiere or to conclude.
During my cleaning process, one of the first things I noticed was that the data scraped included lots of genres where two words were combined into one when they shouldn't be. I filtered out genres that were worded incorrectly and manually added spaces where needed. I also noticed a considerable number of shows that didn't have a genre listed, but upon viewing the Wikipedia page, I found that all of those shows were children's shows. As such, I replaced the null cells to state that the genre was "Children's". Because of the large number of shows that ended up being children's shows, I added another question that I wanted to take away from the analysis: how do children's shows perform compared to those shows that aren't primarily intended for children? The final step of my cleaning was changing the cancellation boolean so that all entries with "1" were true and all entries with "0" were false.
Afterwards, I proceeded to the analysis phase where I wanted to take information away that would help with creating valuable pivot tables, and ultimately, visualizations for my dashboard. With dramas and comedies being very prevalent, I made a column indicating whether a show was a comedy drama, a comedy, a drama, or another category by searching the genre for the words "comedy" and "drama". I used the same process to determine whether or not a show was or was not a children's show. Finally, for each premiere and finale date, I broke the dates down by month, by quarter, and by year.